Update syllabus.md
Added explicit statement about consequence of missing 8 or more days, i.e. an F in the course. Added a statement about use of AI in this course. Added a statement about the closed book nature of the written portion of the exams as well as resources that can be used for the programming portion.